fix(config): address review feedback
