+++ src/app/tests/mod.rs
@@ -1,2 +1,1 @@
-mod config_init;
 mod runtime_worker;
